main/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.6
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-milvus>=0.1.0

# Vector database
pymilvus>=2.3.0

# Environment management
python-dotenv>=1.0.1

# Type hints
typing-extensions>=4.7.0

# Development dependencies
mypy>=1.11.1
ruff>=0.6.1
pytest>=8.3.5
anyio>=4.7.0
langgraph-cli[inmem]>=0.2.8

# Additional dependencies that might be needed
pydantic>=2.0.0
httpx>=0.25.0

# PDF processing
PyMuPDF>=1.23.0

# Document processing
python-docx>=1.1.0

# Text processing
langchain-text-splitters>=0.1.0

# WhatsApp Integration
flask>=2.3.0
twilio>=8.2.0
google-generativeai>=0.3.0

# MCP Integration
mcp>=1.0.0

# Office Word MCP Server for document editing
office-word-mcp-server>=1.1.10

# Async event loop handling
nest-asyncio>=1.5.6
